Implemented negative offset for slice
Changed the slice methods on Series, DataFrame and LazyFrame.
When a negative offset is specified the offset will be counted
from the end of the array.

+++ polars/polars-lazy/src/physical_plan/executors.rs
@@ -805,7 +805,7 @@ impl Executor for StackExec {
 
 pub struct SliceExec {
     pub input: Box<dyn Executor>,
-    pub offset: usize,
+    pub offset: i64,
     pub len: usize,
 }
